moved abstractions of infrastructure to Core project to avoid violation of dependency invertion principle
