+++ CoffeeMachine.Core/Application/Queries/Handlers/GetBeverageStatusQueryHandler.cs
@@ -1,8 +1,8 @@
-﻿using CoffeeMachine.Core.Application.Mediator.Interfaces;
+﻿using CoffeeMachine.Core.Abstractions.Logging;
+using CoffeeMachine.Core.Application.Mediator.Interfaces;
 using CoffeeMachine.Core.Application.Models;
 using CoffeeMachine.Core.Application.Services;
 using CoffeeMachine.Core.Common.Results;
-using CoffeeMachine.Infrastructure.Abstractions;
 
 namespace CoffeeMachine.Core.Application.Queries;
