+++ CoffeeMachine.Core/Application/Services/InventoryService.cs
@@ -1,7 +1,7 @@
-﻿using CoffeeMachine.Core.Application.Models;
+﻿using CoffeeMachine.Core.Abstractions.Logging;
+using CoffeeMachine.Core.Application.Models;
 using CoffeeMachine.Core.Common.Results;
 using CoffeeMachine.Core.Domain.Enums;
-using CoffeeMachine.Infrastructure.Abstractions;
 
 namespace CoffeeMachine.Core.Application.Services;
