+++ CoffeeMachine.Core/CoffeeMachine.Core.csproj
@@ -6,8 +6,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\CoffeeMachine.Infrastructure\CoffeeMachine.Infrastructure.csproj" />
-  </ItemGroup>
-
 </Project>
