+++ CoffeeMachine.Infrastructure/CoffeeMachine.Infrastructure.csproj
@@ -6,4 +6,12 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+	<ItemGroup>
+		<ProjectReference Include="..\CoffeeMachine.Core\CoffeeMachine.Core.csproj" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <Folder Include="Services\Logging\" />
+	</ItemGroup>
+
 </Project>
