+++ CoffeeMachine.UI/App.xaml.cs
@@ -2,8 +2,6 @@
 using Microsoft.UI.Xaml;
 using CoffeeMachine.Core.Application.Mediator.Interfaces;
 using CoffeeMachine.Core.Common.Messages;
-using CoffeeMachine.Infrastructure.Abstractions;
-using CoffeeMachine.Infrastructure.Services.Logging;
 using Microsoft.Extensions.DependencyInjection;
 using CoffeeMachine.Core.Application.Mediator;
 using Microsoft.Extensions.Hosting;
@@ -12,6 +10,8 @@ using CoffeeMachine.Core.Application.Queries;
 using CoffeeMachine.UI.ViewModels;
 using CoffeeMachine.Core.Application.Services;
 using Microsoft.UI.Dispatching;
+using CoffeeMachine.Core.Abstractions.Logging;
+using CoffeeMachine.Infrastructure.Services.Logging;
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
