+++ CoffeeMachine.UI/ViewModels/OrderStatusViewModel.cs
@@ -3,6 +3,7 @@ using System.Reactive.Concurrency;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Input;
+using CoffeeMachine.Core.Abstractions.Logging;
 using CoffeeMachine.Core.Application.Commands;
 using CoffeeMachine.Core.Application.Mediator.Interfaces;
 using CoffeeMachine.Core.Application.Models;
@@ -10,7 +11,6 @@ using CoffeeMachine.Core.Application.Queries;
 using CoffeeMachine.Core.Common;
 using CoffeeMachine.Core.Common.Messages;
 using CoffeeMachine.Core.Common.Results;
-using CoffeeMachine.Infrastructure.Abstractions;
 using CommunityToolkit.Mvvm;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
